# Conflicts:
#	pom.xml
#	src/main/java/com/tarzan/cms/common/config/MybatisPlusConfig.java
#	src/main/java/com/tarzan/cms/module/blog/controller/BlogApiController.java
#	src/main/java/com/tarzan/cms/module/blog/controller/BlogWebController.java
#	src/main/java/com/tarzan/cms/modules/admin/controller/common/BackupController.java
#	src/main/java/com/tarzan/cms/modules/admin/controller/sys/MenuController.java
#	src/main/java/com/tarzan/cms/modules/admin/service/biz/ArticleLookService.java
#	src/main/java/com/tarzan/cms/modules/admin/service/biz/ArticleService.java
#	src/main/java/com/tarzan/cms/modules/admin/service/biz/CategoryService.java
#	src/main/java/com/tarzan/cms/modules/admin/service/biz/ThemeService.java
#	src/main/java/com/tarzan/cms/modules/admin/service/common/CommonDataService.java
#	src/main/java/com/tarzan/cms/modules/blog/controller/BlogApiController.java
#	src/main/java/com/tarzan/cms/modules/blog/controller/BlogWebController.java
#	src/main/java/com/tarzan/cms/modules/foreground/CmsApiController.java
#	src/main/java/com/tarzan/cms/modules/foreground/CmsWebController.java
#	src/main/resources/application.yml
#	src/main/resources/mapper/sys/MenuMapper.xml

+++ src/main/java/com/tarzan/cms/modules/admin/model/biz/Article.java
@@ -37,13 +37,14 @@ public class Article extends BaseVo {
     private Integer original;
     private String description;
     private String keywords;
+    //是否开启评论
     private Integer comment;
     @TableField(exist = false)
-    private Integer lookCount = 0;
+    private Long lookCount = 0L;
     @TableField(exist = false)
-    private Integer commentCount = 0;
+    private Long commentCount = 0L;
     @TableField(exist = false)
-    private Integer loveCount = 0;
+    private Long loveCount = 0L;
     @TableField(exist = false)
     private Integer newFlag = 0;
     @TableField(exist = false)
